requirements.txt
Django==3.2.6
django-ckeditor==6.1.0     
django-crispy-forms==1.12.0
Pillow==8.3.1
django-environ==0.4.5